回退协商代码
# message auto-generated for no-merge-commit merge:
merge 1111zzx into master

回退协商代码


Match-id-45aa2b093c20771de8c4f427aa53c595f551f0bc

+++ src/flow/slice.h
@@ -49,7 +49,7 @@ public:
         return mFlowType;
     }
 
-    inline uint32_t GetLength() const
+    inline uint64_t GetLength() const
     {
         return mLength;
     }
@@ -93,7 +93,7 @@ protected:
 private:
     uint32_t mdataCrc{ 0 };
     FlowType mFlowType{ FLOW_MEMORY };
-    uint32_t mLength{ 0 };
+    uint64_t mLength{ 0 };
     std::vector<FlowAddr> mAddrs;
 };
